Move crown trait functions for crown_qm and crown_z_max into flora module
